Pass current position

+++ sphinx/screens-detail/video-screen/video-screen/src/main/java/chat/sphinx/video_screen/navigation/VideoScreenNavigator.kt
@@ -21,7 +21,7 @@ abstract class VideoScreenNavigator(
         messageId: MessageId,
         videoFilepath: String?,
         feedUrl: FeedUrl? = null,
-        currentTime: Long = 0L
+        currentTime: Int = 0
     ) {
         navigationDriver.submitNavigationRequest(
             ToFullscreenVideoActivity(
